+++ routes/api/contacts.js
@@ -1,83 +1,22 @@
 const express = require("express");
-const joi = require("joi");
+const {
+  getAll,
+  getId,
+  add,
+  deleteById,
+  updateById,
+} = require("../../controllers/contacts");
 
 const router = express.Router();
 
-const {
-  listContacts,
-  getContactById,
-  addContact,
-  removeContact,
-  updateContact,
-} = require("../../models/contacts");
+router.get("/", getAll);
 
-const addShema = joi.object({
-  name: joi.string().required(),
-  email: joi.string().required(),
-  phone: joi.string().required(),
-});
+router.get("/:contactId", getId);
 
-const { httpError } = require("../../utils/httpError");
+router.post("/", add);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await listContacts();
+router.delete("/:contactId", deleteById);
 
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await getContactById(contactId);
-
-    if (!result) throw httpError(404, "Not found");
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addShema.validate(req.body);
-    if (error) throw httpError(400, "missing required name field");
-
-    const result = await addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await removeContact(contactId);
-
-    if (!result) throw httpError(404, "Not found");
-    res.status(200).json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = addShema.validate(req.body);
-    if (error) throw httpError(400, "missing fields");
-
-    const { contactId } = req.params;
-    const result = await updateContact(contactId, req.body);
-
-    if (!result) throw httpError(404, "Not found");
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", updateById);
 
 module.exports = router;
